update: sync MhmRdd dev changes and unify workflows
Apply MhmRdd/ReZygisk dev changes (NoHello syscall 244, csoloader
hiding, PLTI, companion library, etc.) and merge ci/cla/trusted_ci
workflows into a single CI workflow triggerable without secrets.

+++ loader/src/ptracer/monitor.c
@@ -3,8 +3,10 @@
 #include <time.h>
 #include <errno.h>
 
+#include <fcntl.h>
 #include <unistd.h>
 #include <sys/epoll.h>
+#include <sys/mount.h>
 #include <sys/signalfd.h>
 #include <sys/socket.h>
 #include <sys/un.h>
@@ -90,6 +92,16 @@ bool monitor_events_register_event(monitor_event_callback_t event_cb, int fd, ui
   };
 
   if (epoll_ctl(monitor_epoll_fd, EPOLL_CTL_ADD, fd, &ev) == -1) {
+    PLOGE("epoll_ctl");
+
+    return false;
+  }
+
+  return true;
+}
+
+bool monitor_events_unregister_event(int fd) {
+  if (epoll_ctl(monitor_epoll_fd, EPOLL_CTL_DEL, fd, NULL) == -1) {
     PLOGE("epoll_ctl");
 
     return false;
@@ -584,13 +596,13 @@ void sigchld_listener_callback() {
     }
 
     if (s != sizeof(sigchld_fdsi)) {
-      LOGE("read %zu != %zu", s, sizeof(sigchld_fdsi));
+      LOGW("read %zu != %zu", s, sizeof(sigchld_fdsi));
 
       continue;
     }
 
     if (sigchld_fdsi.ssi_signo != SIGCHLD) {
-      LOGE("No sigchld received");
+      LOGW("no sigchld received");
 
       continue;
     }
@@ -608,13 +620,13 @@ void sigchld_listener_callback() {
 
           ptrace(PTRACE_GETEVENTMSG, pid, 0, &child_pid);
 
-          LOGV("Forked %ld", child_pid);
+          LOGV("forked %ld", child_pid);
         } else if (STOPPED_WITH(SIGTRAP, PTRACE_EVENT_STOP) && tracing_state == STOPPING) {
           if (ptrace(PTRACE_DETACH, 1, 0, 0) == -1) PLOGE("failed to detach init");
 
           tracing_state = STOPPED;
 
-          LOGI("Stopped tracing init");
+          LOGI("stop tracing init");
 
           continue;
         }
@@ -622,13 +634,13 @@ void sigchld_listener_callback() {
         if (WIFSTOPPED(sigchld_status)) {
           if (WPTEVENT(sigchld_status) == 0) {
             if (WSTOPSIG(sigchld_status) != SIGSTOP && WSTOPSIG(sigchld_status) != SIGTSTP && WSTOPSIG(sigchld_status) != SIGTTIN && WSTOPSIG(sigchld_status) != SIGTTOU) {
-              LOGW("Injecting signal sent to init: %s %d", sigabbrev_np(WSTOPSIG(sigchld_status)), WSTOPSIG(sigchld_status));
+              LOGW("inject signal sent to init: %s %d", sigabbrev_np(WSTOPSIG(sigchld_status)), WSTOPSIG(sigchld_status));
 
               ptrace(PTRACE_CONT, pid, 0, WSTOPSIG(sigchld_status));
 
               continue;
             } else {
-              LOGW("Suppressing stop signal sent to init: %s %d", sigabbrev_np(WSTOPSIG(sigchld_status)), WSTOPSIG(sigchld_status));
+              LOGW("suppress stopping signal sent to init: %s %d", sigabbrev_np(WSTOPSIG(sigchld_status)), WSTOPSIG(sigchld_status));
             }
           }
 
@@ -651,7 +663,7 @@ void sigchld_listener_callback() {
       }
 
       if (state == 0) {
-        LOGV("New process %d attached", pid);
+        LOGV("new process %d attached", pid);
 
         for (size_t i = 0; i < sigchld_process_count; i++) {
           if (sigchld_process[i] != 0) continue;
@@ -694,7 +706,7 @@ void sigchld_listener_callback() {
 
           do {
             if (tracing_state != TRACING) {
-              LOGD("Stopped injecting %d because status is not set to tracing", pid);
+              LOGW("stop injecting %d because not tracing", pid);
 
               break;
             }
@@ -704,20 +716,30 @@ void sigchld_listener_callback() {
             PRE_INJECT_TANGO
 
             if (tracer != NULL) {
-              LOGD("Stopping %d (program: %s, tracer: %s, tango: %s)", pid, program, tracer, is_tango ? "yes" : "no");
+              LOGI("handoff tracer: pid=%d program=%s tracer=%s tango=%s", pid, program, tracer, is_tango ? "yes" : "no");
 
-              kill(pid, SIGSTOP);
-              ptrace(PTRACE_CONT, pid, 0, 0);
-              waitpid(pid, &sigchld_status, __WALL);
+              if (is_tango) {
+                /* INFO: Stopping tango during init causes an unrecoverable SIGSEGV on resume. */
+                /* TODO: Can this be improved? Can we make an injection without time being a factor? */
+                LOGD("tango deferred: detaching %d without stop", pid);
 
-              if (!STOPPED_WITH(SIGSTOP, 0)) {
-                LOGE("Failed to stop process %d", pid);
+                ptrace(PTRACE_DETACH, pid, 0, 0);
+              } else {
+                LOGD("stopping %d", pid);
 
-                break;
+                kill(pid, SIGSTOP);
+                ptrace(PTRACE_CONT, pid, 0, 0);
+                waitpid(pid, &sigchld_status, __WALL);
+
+                if (!STOPPED_WITH(SIGSTOP, 0)) {
+                  LOGW("handoff: pid %d did not stop as expected", pid);
+
+                  break;
+                }
+
+                LOGD("detaching %d", pid);
+                ptrace(PTRACE_DETACH, pid, 0, SIGSTOP);
               }
-
-              LOGD("Detaching %d", pid);
-              ptrace(PTRACE_DETACH, pid, 0, SIGSTOP);
 
               {
                 sigchld_status = 0;
@@ -744,12 +766,12 @@ void sigchld_listener_callback() {
                     }
                   }
 
-                  PLOGE("exec");
+                  PLOGE("failed to exec, kill");
 
                   kill(pid, SIGKILL);
                   exit(1);
                 } else if (p == -1) {
-                  PLOGE("fork");
+                  PLOGE("failed to fork, kill");
 
                   kill(pid, SIGKILL);
                 }
